Changed Calendar added User for favorites
Changed the calendar stuff. Removed everything from the old calendar and started over with a fragment. Added a new fragment with a viewModel to display a webview to use the calendar from the website.
Added User in the data package to keep the firebase User and the favorites in one location. The User class has access to the database at the user's branch. STILL A WIP
Fixed the nav header to use the appropriate @strings/ Strings