Favorite Teams
It now successfully takes the favorite teams from the database AND shows them prechecked using the fancy star checkbox button. And they are not checked if they are not in the favorites list. It can add and remove the sport to the favorite list, both within the User class AND the database. Have a public static User in MainActivity to keep a global User so that the favorites list can be tracked in only one place.